+++ apps/admin/src/types/api.d.ts
@@ -1,3 +1,2 @@
-/// <reference types="@workspace/ui/services/gateway/typings" />
 /// <reference types="@workspace/ui/services/admin/typings" />
 /// <reference types="@workspace/ui/services/common/typings" />